Parte Grafica/requirements.txt
altgraph==0.17.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comtypes==1.3.0
contourpy==1.2.0
cycler==0.12.1
distlib==0.3.8
et-xmlfile==1.1.0
ezdxf==1.1.4
fastkml==0.12
ffmpeg==1.4
filelock==3.13.1
fonttools==4.48.1
fsspec==2024.3.1
idna==3.6
Jinja2==3.1.3
jiwer==3.0.3
kiwisolver==1.4.5
llvmlite==0.42.0
MarkupSafe==2.1.5
matplotlib==3.8.3
more-itertools==10.2.0
mpmath==1.3.0
networkx==3.2.1
numba==0.59.1
numpy==1.26.3
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opencv-python==4.9.0.80
openpyxl==3.1.2
packaging==24.0
pandas==2.2.0
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.2.0
pyautocad==0.2.0
pycparser==2.21
pygeoif==0.7
pyinstaller==6.6.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
pywin32-ctypes==0.2.2
rapidfuzz==3.7.0
regex==2023.12.25
requests==2.31.0
setuptools==69.5.1
six==1.16.0
sympy==1.12
tiktoken==0.6.0
tkinterdnd2==0.3.0
torch==2.2.2
tqdm==4.66.2
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.1
virtualenv==20.25.0
xlrd==2.0.1
